+++ servant-server/src/Servant/Server/Internal/Authentication.hs
@@ -45,7 +45,6 @@ import           Servant.API.Authentication       (AuthPolicy (Strict, Lax),
                                                    BasicAuth (BasicAuth),
                                                    DigestAuth (..), Algorithm (..))
 
-
 -- | Class to represent the ability to extract authentication-related
 -- data from a 'Request' object.
 class AuthData a where
@@ -118,6 +117,7 @@ basicAuthLax :: KnownSymbol realm
              -> subserver
              -> AuthProtected (BasicAuth realm) usr subserver 'Lax
 basicAuthLax = laxProtect
+<<<<<<< HEAD
 
 
 parseAlgorithm :: B.ByteString -> Maybe Algorithm
@@ -194,3 +194,5 @@ digestAuthHandlers = AuthHandlers onMissingAuthData (const onMissingAuthData)
       return $ responseBuilder status401 [("WWW-Authenticate", headerBytes)] mempty
 
 digestAuthStrict ha1 lookup subserver = strictProtect (digestAuthCheck ha1 lookup) digestAuthHandlers subserver
+=======
+>>>>>>> 8c9d0f2448e36d35ec6b09a0cd3bb844997cb407